requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyPDF2==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.6
requests==2.31.0
beautifulsoup4==4.12.2